Cleaned code and separated middleware
> created authMiddleweare.js
