+++ user_service.js
@@ -2,25 +2,10 @@ const jwt = require("jsonwebtoken");
 const express = require('express');
 const app = express();
 const port = 3002;
+const verifyToken = require('./middleware/authMiddleware');
+
 
 app.use(express.json());
-
-const JWT_SECRET = 'yourSecretKey'; 
-
-function verifyToken(req, res, next) {
-    const token = req.headers['authorization'];
-    if (!token) {
-        return res.status(403).json({ message: 'No token provided' });
-    }
-    const bearerToken = token.split(' ')[1]; // Extract the token
-    jwt.verify(bearerToken, JWT_SECRET, (err, decoded) => {
-        if (err) {
-            return res.status(403).json({ message: 'Failed to authenticate token' });
-        }
-        req.user = decoded; // Attach decoded token info (like user id, role) to req
-        next(); // Proceed to the next middleware or route
-    });
-}
 
 const getUser = async(username) => {
     return {id: 123, password: "12345", username, role:'customer' };
